+++ examples/resources/netapp-ontap_protocols_cifs_local_user/resource.tf
@@ -1,0 +1,7 @@
+resource "netapp-ontap_protocols_cifs_local_user_resource" "protocols_cifs_local_user" {
+  # required to know which system to interface with
+  cx_profile_name = "cluster4"
+  name = "testme"
+  svm_name = "ansibleSVM"
+  password = "netapp123!"
+}